refactor: organize codes by funcionality
