+++ src/apis/mod.rs
@@ -1,0 +1,5 @@
+pub mod mirroring;
+#[cfg(feature = "openid")]
+pub mod openid;
+pub mod registry;
+pub mod user;
