+++ src/apis/registry/delete.rs
