+++ src/apis/registry/put.rs
