+++ test-awrp/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "test-awrp"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tokio = { version = "1.1", features = ["macros", "rt-multi-thread", "fs", "io-util"] }
+warp = "0.3"
+serde_json = "1"
